# docs/backend/openapi.v1.yaml
...
openapi: 3.1.0
info:
  title: LabelScan API
  version: v1
  description: Generated inventory of the currently implemented LabelScan API. Production disables the
    interactive OpenAPI and Swagger endpoints; this reviewed artifact defines pentest scope.
paths:
...
  /v1/me:
    get:
      summary: Get the current user and complete access context
      description: Returns role-derived capabilities plus authorized stores and detailed business portals.
        Super-administrators see the organization, administrators see their owned store portals, and managers
        are limited to their active assignment.
      operationId: get_me_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
      security:
      - bearerAuth: []
  /v1/stores/{store_id}/portals:
    get:
      summary: List the business portals visible for a store
      description: Super-administrators see every portal of an organization-owned store. Administrators
        see portals of stores they own; managers see only their active assignment. An invisible or foreign
        store is reported as not found.
      operationId: list_store_portals_v1_stores__store_id__portals_get
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Store Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalResponse'
                title: Response List Store Portals V1 Stores  Store Id  Portals Get
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
    put:
      summary: Soft-activate or deactivate a store portal
      description: Admin and super-admin only. The requested state is idempotent and never deletes a portal.
        A real state transition is audited atomically; deactivation revokes sessions belonging to users
        assigned to that portal.
      operationId: set_store_portal_active_v1_stores__store_id__portals_put
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Store Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPortalActiveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/me/password:
    post:
      summary: Change My Password
      operationId: change_my_password_v1_me_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/admins:
    get:
      summary: List Admins
      operationId: list_admins_v1_admins_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response List Admins V1 Admins Get
      security:
      - bearerAuth: []
    post:
      summary: Create Admin
      operationId: create_admin_v1_admins_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdentityRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/admins/{user_id}:
    delete:
      summary: Delete Admin
      operationId: delete_admin_v1_admins__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/managers:
    get:
      summary: List Managers
      operationId: list_managers_v1_managers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response List Managers V1 Managers Get
      security:
      - bearerAuth: []
    post:
      summary: Create Manager
      operationId: create_manager_v1_managers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagerRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/managers/{user_id}/portals:
    patch:
      summary: Change Manager Portals
      operationId: change_manager_portals_v1_managers__user_id__portals_patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagerAssignmentsRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
  /v1/managers/{user_id}:
    patch:
      summary: Set Manager Active
      operationId: set_manager_active_v1_managers__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
    delete:
      summary: Delete a manager from active administration
      description: Soft-deletes the account and revokes its access while preserving the identity row used
        by historical labels. The username becomes reusable.
      operationId: delete_manager_v1_managers__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - bearerAuth: []
...
components:
  schemas:
    ActiveRequest:
      properties:
        active:
          type: boolean
          title: Active
      additionalProperties: false
      type: object
      required:
      - active
      title: ActiveRequest
...
    ChangePasswordRequest:
      properties:
        current_password:
          type: string
          maxLength: 128
          minLength: 1
          title: Current Password
        new_password:
          type: string
          maxLength: 128
          minLength: 12
          title: New Password
      additionalProperties: false
      type: object
      required:
      - current_password
      - new_password
      title: ChangePasswordRequest
...
    CreateIdentityRequest:
      properties:
        username:
          type: string
          maxLength: 254
          minLength: 1
          title: Username
        display_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Display Name
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: CreateIdentityRequest
    CreateManagerRequest:
      properties:
        username:
          type: string
          maxLength: 254
          minLength: 1
          title: Username
        display_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Display Name
        password:
          type: string
          maxLength: 128
          minLength: 12
          title: Password
        business_portal_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 1
          minItems: 1
          title: Business Portal Ids
      additionalProperties: false
      type: object
      required:
      - username
      - password
      - business_portal_ids
      title: CreateManagerRequest
...
    ManagerAssignmentsRequest:
      properties:
        business_portal_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 1
          minItems: 1
          title: Business Portal Ids
      additionalProperties: false
      type: object
      required:
      - business_portal_ids
      title: ManagerAssignmentsRequest
    MeResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        capabilities:
          items:
            type: string
          type: array
          title: Capabilities
          description: Canonical capabilities granted by the persisted user role.
        scopes:
          items:
            type: string
          type: array
          title: Scopes
          description: OAuth-style scope names; currently identical to capabilities.
        stores:
          items:
            $ref: '#/components/schemas/labelscan__contexts__identity__adapters__http__access_router__StoreResponse'
          type: array
          title: Stores
          description: Stores visible to the current role and portal assignments.
        business_portals:
          items:
            $ref: '#/components/schemas/PortalResponse'
          type: array
          title: Business Portals
          description: Business portals visible to the current role and assignments.
      type: object
      required:
      - user
      - capabilities
      - scopes
      - stores
      - business_portals
      title: MeResponse
...
    PortalResponse:
      properties:
        id:
          type: string
          title: Id
        store_id:
          type: string
          title: Store Id
        store_code:
          type: string
          title: Store Code
        store_name:
          type: string
          title: Store Name
        profession_code:
          type: string
          title: Profession Code
        profession_name:
          type: string
          title: Profession Name
        name:
          type: string
          title: Name
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - store_id
      - store_code
      - store_name
      - profession_code
      - profession_name
      - name
      - active
      title: PortalResponse
...
    SetPortalActiveRequest:
      properties:
        portal_id:
          type: string
          format: uuid
          title: Portal Id
          description: Portal belonging to the store identified by the URL.
        active:
          type: boolean
          title: Active
          description: Desired soft-activation state. Repeating the same PUT is a no-op.
      additionalProperties: false
      type: object
      required:
      - portal_id
      - active
      title: SetPortalActiveRequest
...
    UserResponse:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        role:
          type: string
          title: Role
        active:
          type: boolean
          title: Active
        organization_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Organization Id
        store_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Store Id
        store_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Store Code
        business_portal_ids:
          items:
            type: string
          type: array
          title: Business Portal Ids
        created_by:
          type: string
          title: Created By
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
      - id
      - username
      - display_name
      - role
      - active
      - organization_id
      - store_id
      - store_code
      - business_portal_ids
      - created_by
      - created_at
      - updated_at
      title: UserResponse
...
    labelscan__contexts__identity__adapters__http__access_router__StoreResponse:
      properties:
        id:
          type: string
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - code
      - name
      - active
      title: StoreResponse
...
    Problem:
      type: object
      description: Sanitized RFC 9457-style error envelope.
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - timestamp
      - retriable
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        error_code:
          type: string
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
        timestamp:
          type: string
          format: date-time
        retriable:
          type: boolean
        errors:
          type: array
          items:
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute JWT with jti, sid, issuer, audience, tenant, role, and scopes
servers:
- url: /
  description: same-origin deployment